feat: Allow for populating ACS during DB migration
The `--populate-acs` flag of the `migrate` sub-command loads all the
ACS documents into the database. This makes deployment much easier.
