+++ acs/acs.go
@@ -1,0 +1,6 @@
+package acs
+
+import "embed"
+
+//go:embed *.json
+var Files embed.FS
